@@ -19,11 +19,10 @@
     "babel-jest": "^5.3.0",
     "babel-loader": "^5.3.2",
     "jest-cli": "^0.5.8",
-    "jsx-loader": "^0.13.2",
     "react-addons": "^0.9.0",
     "webpack": "^1.12.2"
   },
   "dependencies": {
-    "react": "^0.13.3"
+    "react": "^0.14"
   }
 }
