@@ -1,8 +1,8 @@
 'use restrict';
 
-var React = require('react');
+import React from 'react';
 
-var Tabs = require('./tabs.jsx');
+import Tabs from './tabs';
 
 var tabs2 = [
   { name: 'Tab1', description: 'esto es tab1' },
@@ -10,27 +10,26 @@ var tabs2 = [
   { name: 'Tab3', description: 'esto es tab3' },
 ];
 
-var TabSystem2 = React.createClass({
-
-  getInitialState () {
-    return {
-      currentTab: 0
-    }
-  },
+class TabSystem2 extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {currentTab: 0};
+  }
 
   handleClick (currentTab) {
     this.setState({ currentTab })
-  },
+  }
 
   render () {
     return (
       <div>
         <h1>TAB SYSTEM 2</h1>
-        <Tabs tabs = {tabs2} currentTab = {this.state.currentTab} handleClick = {this.handleClick}/>
+        <Tabs tabs = {tabs2} currentTab = {this.state.currentTab} handleClick = {this.handleClick.bind(this)}/>
         <hr />
       </div>
     )
   }
-});
+};
 
-module.exports = TabSystem2;
+
+export default TabSystem2;
